Split Pulse page presentational components
